backend/requirements.txt
aiofiles==0.7.0
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
black==24.8.0
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.112.1
flake8==7.1.1
frozenlist==1.4.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
openai==0.27.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
py==1.11.0
pycodestyle==2.12.1
pydantic==1.8.2
pyflakes==3.2.0
pytest==6.2.5
python-dotenv==1.0.1
python-multipart==0.0.5
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
toml==0.10.2
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.15.0
yarl==1.9.4
